renamings
renamed folder to index-fund
remove export from GuardPrompt
