+++ src/pages/Admin/Core/Templates/index-fund/Alliance/schema.ts
