+++ src/pages/Admin/Core/Templates/index-fund/Config/schema.ts
