+++ src/pages/Admin/Core/Templates/index-fund/RemoveFund/index.tsx
