+++ src/pages/Admin/Core/Templates/index-fund/RemoveFund/schema.ts
