+++ src/pages/Admin/Core/Templates/index-fund/RemoveFund/useDestroyFund.ts
