+++ src/pages/Admin/Core/Templates/index.tsx
@@ -3,12 +3,12 @@ import { templateRoutes as routes } from "../../constants";
 import Config from "../../templates/cw3/Config";
 import FundSender from "../../templates/cw3/FundSender";
 import Members from "../../templates/cw4/Members";
-import Alliance from "./indexfund/Alliance";
-import FundConfig from "./indexfund/Config";
-import CreateFund from "./indexfund/CreateFund";
-import IndexFundOwner from "./indexfund/IndexFundOwner";
-import FundMembers from "./indexfund/Members";
-import RemoveFund from "./indexfund/RemoveFund";
+import Alliance from "./index-fund/Alliance";
+import FundConfig from "./index-fund/Config";
+import CreateFund from "./index-fund/CreateFund";
+import IndexFundOwner from "./index-fund/IndexFundOwner";
+import FundMembers from "./index-fund/Members";
+import RemoveFund from "./index-fund/RemoveFund";
 import RegistrarConfig from "./registrar/Config";
 import EndowmentStatus from "./registrar/EndowmentStatus";
 import RegistrarOwner from "./registrar/Owner";
