+++ src/pages/Admin/Guard.tsx
@@ -48,7 +48,7 @@ export const useAdminResources = (): AdminResources => {
   return val;
 };
 
-export function GuardPrompt(props: { message: string; showLoader?: true }) {
+function GuardPrompt(props: { message: string; showLoader?: true }) {
   return (
     <div className="place-self-center grid content-center justify-items-center bg-white-grey text-angel-grey min-h-[15rem] w-full max-w-sm p-4 rounded-md shadow-lg">
       {props.showLoader ? (
